requirement.txt
absl-py                 0.12.0
alembic                 1.6.2
apturl                  0.5.2
argon2-cffi             20.1.0
asn1crypto              0.24.0
astor                   0.8.1
async-generator         1.10
attrs                   21.2.0
autopep8                1.6.0
backcall                0.2.0
black                   22.8.0
bleach                  1.5.0
Brlapi                  0.6.6
cached-property         1.5.2
certifi                 2018.1.18
certipy                 0.1.3
cffi                    1.14.5
chardet                 3.0.4
click                   8.0.4
cmake                   3.22.0
command-not-found       0.3
cryptography            3.4.7
cupshelpers             1.0
cycler                  0.10.0
dataclasses             0.8
decorator               4.4.2
defer                   1.0.6
defusedxml              0.7.1
distro-info             0.18ubuntu0.18.04.1
entrypoints             0.3
future                  0.18.2
gast                    0.4.0
google-pasta            0.2.0
greenlet                1.1.0
grpcio                  1.37.1
h5py                    3.1.0
html5lib                0.9999999
httplib2                0.9.2
idna                    2.6
imageio                 2.9.0
importlib-metadata      4.0.1
importlib-resources     5.4.0
ipykernel               5.5.5
ipython                 7.16.1
ipython-genutils        0.2.0
ipywidgets              7.6.3
jedi                    0.18.0
Jinja2                  3.0.0
joblib                  1.1.0
jsonschema              3.2.0
jupyter                 1.0.0
jupyter-client          6.2.0
jupyter-console         6.4.0
jupyter-core            4.7.1
jupyter-telemetry       0.1.0
jupyterlab-pygments     0.1.2
jupyterlab-widgets      1.0.0
Keras                   2.2.4
Keras-Applications      1.0.8
Keras-Preprocessing     1.1.2
keyring                 10.6.0
keyrings.alt            3.0
kiwisolver              1.3.1
language-selector       0.1
launchpadlib            1.10.6
lazr.restfulclient      0.13.5
lazr.uri                1.0.3
lmdb                    1.2.1
louis                   3.5.0
macaroonbakery          1.1.3
Mako                    1.1.4
Markdown                3.3.4
MarkupSafe              2.0.0
matplotlib              3.3.4
mistune                 0.8.4
mock                    4.0.3
mypy-extensions         0.4.3
nbclient                0.5.3
nbconvert               6.0.7
nbformat                5.1.3
nest-asyncio            1.5.1
netifaces               0.10.4
networkx                2.5.1
notebook                6.3.0
numpy                   1.19.5
oauth                   1.0.1
oauthlib                3.1.0
olefile                 0.45.1
opencv-contrib-python   4.5.2.52
opencv-python           4.1.2.30
pamela                  1.0.0
pandas                  1.1.5
pandocfilters           1.4.3
parso                   0.8.2
pathspec                0.9.0
pexpect                 4.2.1
pickleshare             0.7.5
Pillow                  5.4.1
pip                     21.3.1
platformdirs            2.4.0
prometheus-client       0.10.1
prompt-toolkit          3.0.18
protobuf                3.17.0
ptyprocess              0.7.0
pycairo                 1.16.2
pycodestyle             2.8.0
pycparser               2.20
pycrypto                2.6.1
pycups                  1.9.73
pyface                  7.4.1
Pygments                2.9.0
PyGObject               3.26.1
pymacaroons             0.13.0
PyNaCl                  1.1.2
pyOpenSSL               20.0.1
pyparsing               2.4.7
PyQt5                   5.15.4
PyQt5-Qt5               5.15.2
PyQt5-sip               12.8.1
pyRFC3339               1.0
pyrsistent              0.17.3
python-apt              1.6.5+ubuntu0.7
python-dateutil         2.8.1
python-debian           0.1.32
python-editor           1.0.4
python-json-logger      2.0.1
pytz                    2018.3
PyWavelets              1.1.1
pyxdg                   0.25
PyYAML                  3.12
pyzmq                   22.0.3
qtconsole               5.1.0
QtPy                    1.9.0
reportlab               3.4.0
requests                2.18.4
requests-unixsocket     0.1.5
robustml                0.0.3
ruamel.yaml             0.17.4
ruamel.yaml.clib        0.2.2
scikit-image            0.17.2
scikit-learn            0.24.2
scipy                   1.5.4
screen-resolution-extra 0.0.0
seaborn                 0.11.1
SecretStorage           2.3.1
semantic-version        2.8.5
Send2Trash              1.5.0
setuptools              56.2.0
setuptools-rust         0.12.1
simplejson              3.13.2
six                     1.16.0
sklearn                 0.0
SQLAlchemy              1.4.15
ssh-import-id           5.7
system-service          0.3
systemd-python          234
tensorboard             1.13.1
tensorboardX            2.4
tensorflow-estimator    1.13.0
tensorflow-gpu          1.13.1
tensorflow-hub          0.6.0
termcolor               1.1.0
terminado               0.9.5
testpath                0.4.4
threadpoolctl           3.1.0
tifffile                2020.9.3
toml                    0.10.2
tomli                   1.2.3
torch                   1.6.0+cu101
torchvision             0.7.0+cu101
tornado                 6.1
tqdm                    4.62.2
traitlets               4.3.3
traits                  6.3.2
turingas                0.1
typed-ast               1.5.4
typing_extensions       4.1.1
ubuntu-advantage-tools  8001
ubuntu-drivers-common   0.0.0
ufw                     0.36
unattended-upgrades     0.1
urllib3                 1.22
usb-creator             0.3.3
wadllib                 1.3.2
wcwidth                 0.2.5
Werkzeug                1.0.1
wheel                   0.30.0
widgetsnbextension      3.5.1
wrapt                   1.12.1
xkit                    0.0.0
yapf                    0.32.0
zipp                    3.4.1
zope.interface          4.3.2